requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch>=1.10.0
torchvision>=0.11.0
pillow==8.3.1
python-multipart==0.0.5
pyjwt==2.1.0
passlib==1.7.4
pytest==6.2.5
python-jose==3.3.0
bcrypt==3.2.0 